[components] split collectors up
