+++ lib/newrelic_aws/collectors/rds.rb
@@ -1,0 +1,5 @@
+module NewRelicAWS
+  module Collectors
+    class RDS < Base; end
+  end
+end
